+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - [fix:fix the ratelimit bug for hoxton](https://github.com/Tencent/spring-cloud-tencent/pull/1301)
 - [feat:upgrade jacoco version.](https://github.com/Tencent/spring-cloud-tencent/pull/1306)
 - [fix:fix no registry when lossless is disabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1313)
+- [fix: memory not released while using wildcard api call with circuitbreaker enabled](https://github.com/Tencent/spring-cloud-tencent/pull/1335)